Combined /traits and /axis into /util
