--- src/axis.rs
@@ -1,11 +1,0 @@
-mod indexer_impl;
-pub(crate) mod indexer;
-
-pub mod index;
-#[macro_use]
-pub mod slice_index;
-
-pub mod axes_traits;
-pub use axes_traits::*;
-
-pub struct Axis(pub isize);
