+++ src/lib.rs
@@ -1,14 +1,11 @@
-pub mod axis;
-pub use axis::*;
-
 pub mod tensor;
 pub use tensor::*;
 
 pub mod linalg;
 pub use linalg::*;
 
-mod traits;
-mod macros;
+pub mod util;
+pub use util::*;
 
 #[cfg(use_apple_accelerate)]
 mod accelerate;
