+++ src/tensor/constructors.rs
@@ -2,10 +2,10 @@ use crate::dtype::NumericDataType;
 use crate::tensor::dtype::RawDataType;
 use crate::tensor::flags::TensorFlags;
 use crate::tensor::Tensor;
-use crate::traits::flatten::Flatten;
-use crate::traits::nested::Nested;
-use crate::traits::shape::Shape;
-use crate::traits::to_vec::ToVec;
+use crate::util::flatten::Flatten;
+use crate::util::nested::Nested;
+use crate::util::shape::Shape;
+use crate::util::to_vec::ToVec;
 use crate::{FloatDataType, TensorMethods};
 use num::NumCast;
 use std::mem::ManuallyDrop;
