+++ src/util/shape.rs
@@ -1,5 +1,5 @@
 use crate::recursive_trait_base_cases;
-use crate::traits::homogenous::Homogenous;
+use crate::util::homogenous::Homogenous;
 
 pub(crate) trait Shape: Homogenous {
     fn shape(&self) -> Vec<usize>;
